Separated the commands of the CLI into several files
It was becoming a huge mess, so now each command is on its own file.
